basic support for monochrome terminals
curses does not automatically emulate colors with reverse text on monochrome terminals.

Scintilla however is based on RGB colors, so theoretically we could get the RGB components
of the foreground and background colors and calculate a contrast value that tells us
wether the text is light-on-dark or dark-on-light.
Instead we implemented a much simpler heuristic: on terminals without color support,
we render text in reverse if the background color is not black (COLOR_BLACK).
This should at least work if you didn't redefine colors with init_color().

This actually makes Scinterm usable in historic terminals like VT100, VT220
or VT240. Perhaps more importantly, $TERM may be set to vt100 or some other
conservative value even on modern serial terminal connections.
Also, there are monochrome variants of most terminals in terminfo.

This introduces a new term_color_attr() which returns either a color
attribute *or* A_REVERSE. This can only be passed to "legacy" attribute
setting functions. However since Scinterm practically restricts us to a palette
of 16 colors, the "legacy" API should be safe to use.